Inconsistent gam path, s/domain/section, no if
I had /bin/ in the suggestion for .bashrc which I then didn't use in the gamselect.sh.
GAM=$HOME/bin/gamadv-xtd3/gam vs GAM=$HOME/gamadv-xtd3/gam
Also the GAMSECTION variable wasn't pulling the value from the same place in the config. Replaced 'domain =' with 'Section:'.
Removed the IF clause in the .bashrc for the initial PS1. Don't know if that can be altered and re-added, but I know most people run shells with colour, so no point in asking.